third_party/blink/web_tests/external/wpt/digital-credentials/get.tentative.https-expected.txt

This is a testharness.js-based test.
[FAIL] navigator.credentials.get() API rejects if there are no providers in cross-origin iframe.
  assert_equals: expected "TypeError" but got "DOMException"
[FAIL] navigator.credentials.get() promise is rejected if called with an aborted controller.
  promise_rejects_dom: function "function() { throw e }" threw object "NotSupportedError: The user agent does not support public key credentials." that is not a DOMException AbortError: property "code" is equal to 9, expected 20
[FAIL] navigator.credentials.get() promise is rejected if called with an aborted controller in same-origin iframe.
  promise_rejects_dom: function "function() { throw e }" threw object "NotSupportedError: The user agent does not support public key credentials." that is not a DOMException AbortError: property "code" is equal to 9, expected 20
[FAIL] navigator.credentials.get() promise is rejected if called with an aborted signal in cross-origin iframe.
  assert_equals: expected "AbortError" but got "NotAllowedError"
[FAIL] navigator.credentials.get() promise is rejected if abort controller is aborted after call to get() in cross-origin iframe.
  assert_equals: expected "AbortError" but got "NotAllowedError"
[FAIL] Mediation is required to get a DigitalCredential.
  assert_unreached: Should have rejected: undefined Reached unreachable code
Harness: the test ran to completion.
